Initialize the dev-suite repo
This commit initializes the repo with a stubbed out ticket tool and the
rustfmt preferences for the repo. The idea is that dev-suite will allow
remote collaboration by giving a lot of the functionality that GitHub
and other services have, but have all of the data live alongside the
repo and it's history. This makes choosing a different service easier
and lets people who don't want to use the service have that option.