Move find_root function into the new shared crate
This cleans up the init function using the modified find_root function
for ticket and moves it into a new shared crate so that other tools that
might be built can use it. This means we can easily find the root of
a git repo no matter where in the repo one is and build paths relative
to it.

Closes #3