Add hooked and empty inited hooks from the tool
This adds the hooked binary to the dev-suite repo as well as a stub for
a program to be used in this workflow! Hooked works by adding the hooks
into the repo and setting them to executable and linking them into the
hooks directory under .git. This means hooks get to travel with the
repo and are source controlled. All a dev needs to do is run the init
command and hooked will symlink them all for them. No need to remember
how ln works. It's all handled for you. Future work will iterate about
what hooks that dev-suite supplies as part of the script. This will
involve configuration files and per repo settings are something that
will need to be thought about.

Closes Issue #2