Fix formatting and add checks to pre-commit hook
This enables a pre-commit script and adding more pedantic checks to the
commit. This means from now on all commits will be in a working state in
the history and this enables us to build directly on master without
worrying about it breaking the build. Where we're going we won't need
feature branches anymore. This also fixes formatting issues that existed
but the GitHub actions would not be able to catch at all.