Add commit message linting hook to the repo
This adds a commit to handle git commit linting to enforce style by not
allowing less than 10 or more than 50 chars for titles and less than or
equal to 72 chars for the body. Chars are measured in number of
graphemes as 50 chars represented in the terminal is what we want to use
not 50 bytes. This will eventually be an installable hook for end users
if they want it.