Create dev-suite tool to orchestrate tooling
The dev-suite tool acts simmilar to rustup in that it's responsible for
keeping the tools up to date, installing the tools, and managing itself.
It also includes an init command to run all the various tools init
commands all at once. Of course we want what tools people use to be
configurable. dev-suite uses dialouger in order to provide a nice text
based menu for things like selecting what tools to use etc. Certain
functions are stubbed out for now, but they will be expanded over time.