Add logging output to hooked
We really should be logging what's going on as each of these tools run.
Before this change hooked just ran without any indication of what was
going on. This change adds logging with info level for the end user by
default, with debug and trace statements while developing the code being
an option as well.