Add logging output to ticket
ticket used to just run without any kind of logging for commands that
weren't just printing tickets to the console or getting any kind of
information as to what was going on. This change adds logging with info
level for the end user by default, with debug and trace statements
while developing the code being an option via the RUST_LOG env var.