Make the pre-commit script pedantic and fix errors
This commit really ups the level and quality of the Rust code by setting
clippy to pedantic mode. It also fixes an issue where bash continued to
run scripts even if something failed with a non-zero exit status. We
also deny all warnings so as to actually fail the builds and the commit
hooks. This should make sure code quality stays at a high level.