Add the ability to add comments to tickets
This commit encompasses quite a few changes to add tickets:

- The `Ticket` struct is updated to properly order comments using a uuid
  v1 and to hold the user name, uuid, and their comment
- Tickets in the repo are updated to accomodate this change to the
  ticket. Despite this being a breaking config change, none of these
  tickets had any comments so it was an easy manual port and the
  migration tool did not need to be updated.
- The TUI was updated to display the tickets a bit better with some
  coloring and now also showing the comments with them

This gets us one step closer to a decent first release for ticket. The
only things that are really left to do are adding the ability to comment
in the tui, listing tickets on the cli, and adding in issue assignees on
both the cli and tui.