Upgrade tui to allow commenting from it
This is a fairly large overhaul of ticket but lays down the last bit of
groundwork needed before an initial release. It handles input to write
comments, refactors the code a bit to be cleaner and less computation
heavy, and also adds instructions on the bottom for how to use the tui.
This should be enough for people to start using it, though obviously
there's more work to go, but it feels more usable than before.