Change ticket tui to fix thread panic on Windows
One of the issues experienced on Windows with the tui for ticket was
that the thread handling events was panicking on close, but this was
only happening on Windows. For some reason it must not have been closing
down the spawned thread properly. To get around this we set up another
channel to send a message to close the thread down and avoid the panic.