Cleanup ds install and format hooked properly
This cleans up the ds install command to use namespaced folders on
digital ocean to hold assets. This will make it easier to install things
in the future if we break it down by version, but really it just makes
it easier to not have to do weird splitting logic to choose the right
tool and download and install it. For some reason hooked was also not
formatted properly when rustfmt was run on OSX so this was fixed to
make the build not break.