Fix the ticket tui so it does not panic
The tui for ticket uses indexing in order to have quicker unchecked
access into data structures for the data needed that gets displayed in
the tui. However, you can't index empty data. We add a check for this in
places that expect data to exist so that nothing will cause a panic. We
also add fixes to handle the fact that you can't divide by zero in our
index modifying logic when pressing the left/right/up/down arrow keys.
With this the tui shouldn't panic anymore and can work in an empty
state.

Closes: 4c37e800-2e38-11ea-b6e0-32f54a3ad7cd 'Having no tickets causes
the TUI to crash'